No need to replace a body part with the exact same body part.
No idea why this was done. Traced back to commit 0d2544c. Started
polling the IMAP server for both the original part and the bodypart.
Previous commit was 6c6db9e, when support for modifying objects
was first introduced (even then, hard to say if this was ultimately
happening since the actions are so spread out among different, deep,
class structures.

Anyway, this seems to work, and fixes an issue that was causing
the data to be decoded into some kind of binary data instead of
the XML data is really is.