Kolab_Storage: Improve cleanup of duplicated objects
In the real world it can happen that two messages
with the some Kolab XML uid appear in an IMAP folder.

This could be due to user error or a Kolab client
that first stores the new version of an object
and then deletes the old message.

The current code removes the Kolab object completely
from the cache when an old duplicate message gets expunged.

Fix it by only removing the object from the cache
if the latest version of the object gets deleted.

The fix is not perfect as it does not cover the case
when the latest version (=highest IMAP uid) gets removed
and the old one remains. That case is more complicated
as we need to update the Horde_History, too.